chore: rename alignment guides component
